+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Do not exit the network task when a draining TcpStream is already closed. 
 - Make log messages for "no current WireGuard session" more user-friendly.
 - Attempt to build binary wheels for `aarch64-unknown-linux-gnu` for Raspberry
   Pi support.